AI_devs4/se01e05/code/pyproject.toml
[project]
name = "hitl-api-explorer"
version = "0.1.0"
description = "HITL API exploration agent with OpenRouter integration"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "openrouter>=0.7.11",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
